FluidVids: Support shadow dom
Now fluid vids will work as expected inside or outside a shadow dom.

Note: This commit does use getRootNode which likely changes the supported envs.

For us, this is not a huge deal and likely could be fixed with a small if check
to see if that function is supported.